Performance++ by replacing buggy onCommitted
webNavigation.onCommitted wakes the extension even if the UrlFilter
forbids it to activate (see http://crbug.com/257851).
Replaced the method with a declarativeWebRequest rule + message
dispatcher.

After this commit, the extension will only wake up if needed.
One minor issue: The button in the CWS does not show up by default
if one directly enters the URL. This cannot be detected through the
DWR, because Chromium does not allow extensions to interact with the
Chrome Web Store.