Unified zip export + pubkey in console, fix build
As of this commit, the Download button at the View source screen and
popup behave identical. Before this change, the download button at the
popup just showed the "CRX" file, with a .zip extension. This commit
strips CRX-specific headers, and presents  "real" zip file for download.

The public key is logged to the console, in case it's needed.

Added chrome-platform-info.js (previous commit) to build.sh...