Add font-size and font-family
On previous versions of Chrome, extension pages use the default serif fonts just like on normal web pages.

Now, extension pages have default operating-system-like styling that sets a default font-family and font-size.
https://codereview.chromium.org/400343002

However, the default font-size is set to 75% on Windows and Mac OS (not sure about Linux or Chrome OS), which is a bit small. This commit bumps the font-size back to 100%.

Also, the font-family is set so that the current stable versions of Chrome can benefit from some aesthetic changes.