crx_url var = the state of the opened file
The use of crx_url as a local and global variable is confusing.
So, use local variables only. `window.crx_url` is still set because
it is convenient to be able to open the devtools and use `copy(crx_url)`
to get the URL from the parameters.
jshint will enforce that crx_url is not accidentally used.