Dynamically change checkbox size on Firefox
The dimensions of a HTML checkbox are fixed and cannot be changed in
Firefox. This patch resizes them via CSS transformations, calculated on
the fly. I resized checkboxes instead of using custom ones so that the
checkboxes adhere as much to the platform conventions as possible, so
that the user can recognize that it is a clickable checkbox.