Split up webstore form
To clarify what it means when the form is submitted.
Now it is obvious: If the CWS form is used, then the extension ID must
be valid. And the URL is only updated if it was a webstore URL.
If the user wishes to open a webstore URL while another non-CWS URL had
been input in the top URL input, then they can click on the new button.