Refactor: allow highlight in a line range
No behavioral change, this is a preparation for improving performance of
automatic highlighting. Currently, highlighting all results for
"!regexp:." in jquery.js leads to a terrible performance in Firefox (and
even worse in Chrome).

The plan is to only highlight results in the currently visible view,
and lazily highlight the others when the lines are moved inside the
view.