Refactor: add willSwitchSourceView call
Add explicit notification for when the viewer switches to another file.
In the past, viewing is as simple as inserting a DOM node in the
document, but with the recent changes (find and all), this is becoming
unsustainable. The new (not yet used) sourceviewhide event can be used
to suspend expensive operations if needed.