Explicitly spell out review status of AMO add-ons
AMO (addons.mozilla.org) can show multiple versions of an add-on.
Previously the context menu said "View extension source", but it
was not obvious that the requested extension source could differ
from the actual extension source.
So "(latest approved version)" is now appended to context menu
items that link to an unspecified version of an add-on.