Use different text highlighting method #37
The previous text highlighting method relies on the assumption that
a text node can be split in smaller parts and still follow the same
word wrapping rules as the original text. This does not always work,
as #37 shows.

This commit changes the implementation. Now instead of delegating the
generation of the outline around the selection to the browser, we will
calculate the bounding boxes and manually render the outlines (via SVG).
The dimensions are based

This current commit fixes #37, but needs some small improvements before
it is ready for release:
- The code assumes that the rendered text is identical to the original
  input text. This is not always the case: browsers seem to strip
  0-bytes and replace \r with \n.
- The highlights need to be updated whenever the container resizes.
  This is already done for window resizing, but also needs to be done
  for when the left panel is toggled/resized.