Improve calculation of bounding box in edge cases
Previously, for search terms that are close to the line boundary (e.g.
"!regexp:\n"), the path around the result could skew diagonally and end
up too much to the left. This seemed to happen more frequently in
Firefox than Chrome.

Besides the improved visual positioning, this also improved the
performance by reducing layout trashing for empty lines. Previously an
empty text node was added to an empty element in order to obtain a
suitable node to feed to Range. Now we manually find non-empty lines if
possible.
(Note that the empty text node did not work as desired, the text node
should be non-empty to get a non-empty ClientRect (e.g. with "\u200b")).