Update for compat with older browsers
The viewer works fully in Firefox 46.
Search does not work in earlier versions due to the absence of
FileReader in web workers.

In Chrome 49, the viewer also works completely.
Earlier versions (tested back until 31) do not support
in-file search due to lack of support for ES6 syntax in
src/search-tools.js

In IE 11, the viewer does not load due to the use of the CustomEvent
constructor.

In Edge 42.17134.1.0, all functionality works.

In Safari 11.1.2, everything works (as of this patch).