Update Prism to fix regexp perf issue (#80)
The version number is still 1.15.0, but this build is based on
https://github.com/PrismJS/prism/tree/439ea1ee9abdb63fb90a1cb61bea5860488f80ae

To verify, clone the Github repo, and start a local file server.
Then open its download.html (at localhost) with the parameters at the
top of src/lib/prettify/prism.js , select the "Development version"
option (opposed to "Minified") and generate the prism.js file via the
"Download js" button.