Restore "Hide this button" option from #41
This basically reverts fec6e7c72fcb1324d48d9475591a6970115fb5a8,
which is a commit that removes the functionality from #41.
The logic looks different since in between that revert and now the logic
got refactored to use the more efficient `page_action.show_matches`.

The functionality was originally removed because Firefox 58 has since
introduced the logic in bugzil.la/1395387.

The functionality is restored because Firefox 89+ has regressed and lost
the ability to hide arbitrary buttons, see bugzil.la/1712556.