Options page: wrap text as work-around
When a long error message is shown, Chrome keeps flashing because it
repeatedly tries to resize the content. This happens repeatedly because
upon resizing the content, the lines may reflow and cause scrollbars to
appear and disappear.

To prevent this from happening, use wrap text where possible instead of
at the (default) word boundary.