Introduce getPlatformInfoAsync
chrome-platform-info.js used to call runtime.getPlatformInfo()
unconditionally from the background, and cache it in localStorage,
for use in popup.js and crxviewer.js (and later restarts of the
background, in bg-contextmenu.js). To prepare for worker-based
extensions, remove this localStorage dependency.

As an alternative to relying on the value to be available synchronously,
refactor the callers to explicitly call getPlatformInfoAsync() before
calling getPlatformInfo() itself (whether directly, or indirectly
through `get_crx_url`). As before, a fallback is used if `get_crx_url`
is called without the completion of the async getPlatformInfo call.
This situation does not happen with the current code, except
theoretically in popup.js, which has been annotated with a comment.