Reliable, cross-browser menu registration
This patch replaces browser-specific separate branches with a new
implementation that is more reliable than the previous implementation.

To prepare for migrating to event pages in Firefox, replace the
unconditional context menu registration with a conditional one.
While `runtime.onInstalled` is recommended, it cannot be used because of
bugs in some cases that results in the loss of context menus.

The same logic is used in Chrome, because Chrome also suffers from lost
menu registrations under some conditions. The current code relies on
incognito-events.js to patch the flaw of `runtime.onInstalled` by firing
it when needed, but I want to remove that file because it relies on APIs
unavailable to service workers.