Remove incognito-events.js + add new onInstalled workaround
The incognito-events.js file offered a hack to make sure that
runtime.onInstalled was fired in incognito mode, when the incognito
profile had started for the first time. The implementation relied
on a cookie, which is not ideal and unavailable to service workers.

Instead of fixing the hack, I moved away from runtime.onInstalled
in a previous commit, and added a new API-specific work-around in
this commit (for the declarativeContent rule registration).

With these fixes, the incognito-events.js file is now obsolete.