Fix contextmenu registration in Firefox
In the last edit to bg-contextmenu.js, I assumed that
contextMenus.update would raise an error for non-existent menu
registrations. This is not the case.

To still detect whether the menu was registered, take advantage of the
implementation detail that the validity of parentId is only checked for
existing menus, i.e. that validation is skipped for non-existing menus.