Fix togglePageAction condition for Firefox < 59
In 2037153e55691dae7dbbb62226275f03350b21e3, I added a condition that
was supposed to represent "Firefox < 59". The condition was mistakenly
inverted, which prevents the pageAction logic from applying immediately.
In Firefox >= 59, the call is a no-op.
In Firefox < 59, the logic should have run, but did not. The impact is
minimal because the same logic runs anyway once `storage.sync.get`
resolves.

So the impact of this issue is minimal.